Renamed ribcage module to main
When this project is installed via volo, volo creates a separate ribcage module
which loads the main module from this project. The entire project is then
placed under `ribcage` directory inside the library directory, so you get an
ugly-looking import like `lib/ribcage/ribcage`. This is a cosmetic change to
avoid the ugliness.